Refactor into modular structure with central config
