+++ src/config/__init__.py
@@ -1,0 +1,6 @@
+"""Configuration helpers for the masking application."""
+
+from .models import Config
+from .loader import load_config, create_engine
+
+__all__ = ["Config", "load_config", "create_engine"]
