+++ src/core/__init__.py
@@ -1,0 +1,5 @@
+"""Core masking engine components."""
+
+from .engine import MaskingEngine
+
+__all__ = ["MaskingEngine"]
