+++ src/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility helpers used across modules."""
+
+from .io import read_yaml
+
+__all__ = ["read_yaml"]
